\name{pbd_travis-package}
\alias{pbd_travis-package}
\alias{pbd_travis}
\docType{package}
\title{
\packageTitle{pbd_travis}
}
\description{
\packageDescription{pbd_travis}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{pbd_travis}
\packageIndices{pbd_travis}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{pbd_travis}

Maintainer: \packageMaintainer{pbd_travis}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
